+++ eras/conway/impl/test/Main.hs
@@ -2,6 +2,7 @@ module Main where
 
 import Test.Cardano.Ledger.Common
 import qualified Test.Cardano.Ledger.Conway.BinarySpec as BinarySpec
+import qualified Test.Cardano.Ledger.Conway.GenesisSpec as GenesisSpec
 import qualified Test.Cardano.Ledger.Conway.RatifySpec as RatifySpec
 
 main :: IO ()
@@ -10,3 +11,4 @@ main =
     describe "Conway" $ do
       BinarySpec.spec
       RatifySpec.spec
+      GenesisSpec.spec